backend/requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
awscli==1.22.87
billiard==3.6.4.0
boto3==1.21.32
botocore==1.24.32
cachetools==5.0.0
celery==5.2.5
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
Deprecated==1.2.13
Django==3.2.6
django-cors-headers==3.11.0
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.12.0
djangorestframework-simplejwt==5.1.0
docutils==0.15.2
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
jmespath==1.0.0
json-tricks==3.15.5
kombu==5.2.4
Markdown==3.3.6
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.28
protobuf==3.20.1
psycopg2==2.9.3
ptvsd==4.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
redis==4.2.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.8.0
